backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
mistralai
